# 7-studierendenverwaltung-openapi/openapi-3-spec.yml
openapi: 3.1.0
info:
  title: API definition for managing students
  version: 1.0.0
  description: A API to manage students, courses, teachers and rooms
servers:
  - url: https://api.studentmanagement.de/v1
    description: Production Environment
  - url: https://api-test.studentmanagement.de/v1
    description: Test Environment
paths:
  /students:
    get:
      tags:
        - students
      summary: Get all students
      responses:
        '200':
          description: Get a list of all students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
    post:
      tags:
        - students
      summary: Create a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '201':
          description: Student has been created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
  /students/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: The student's ID
    get:
      tags:
        - students
      summary: Get a student by his ID
      responses:
        '200':
          description: Student Details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Student has not been found with the given id
    put:
      tags:
        - students
      summary: Update a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '200':
          description: Student has been updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Student has not been found with the given id
    delete:
      tags:
        - student
      summary: Delete a student
      responses:
        '204':
          description: Student has been deleted successfully
        '404':
          description: Student has not been found with the given id
...
components:
  schemas:
    Student:
      type: object
      properties:
        id:
          type: integer
          description: Unique ID of a student
          example: 101
        firstname:
          type: string
          description: First name of a student
          example: Max
        lastname:
          type: string
          description: Last name of a student
          example: Mustermann
        matrikelnummer:
          type: string
          description: Matrikelnummer of a student
          example: "S987654"
        enrolledCourses:
          type: array
          items:
            $ref: '#/components/schemas/CourseSimple'
          description: List of courses the student has enrolled for
      required:
        - id
        - firstname
        - lastname
        - matrikelnummer
    StudentInput:
      type: object
      properties:
        firstname:
          type: string
          description: First name of a student
          example: Erika
        lastname:
          type: string
          description: Last name of a student
          example: Mustermann
        matrikelnummer:
          type: string
          description: Matrikelnummer
          example: "S123456"
      required:
        - firstname
        - lastname
        - matrikelnummer
...
    CourseSimple:
      type: object
      properties:
        id:
          type: integer
          description: Unique ID of the course
          example: 201
        name:
          type: string
          description: Name of the course
          example: Einführung in die Programmierung
      required:
        - id
        - name
...